requirement.txt
requests
beautifulsoup4
faiss-cpu
numpy
sentence-transformers
